remove: detection point by caching mountinfo
This commit removes a detection point by caching mountinfo. Native Test Futile Hide (01) is a detection which is able to detect "/proc/self/mountinfo" reads in unshare, due to that it is not possible to read that file in the unshare syscall. To bypass that we cache the Magisk/KSU/APatch mounts and umount all of them in unshare.