update: translations; add: new translations; remove: incomplete translations
This commit updates and adds new translations based on contributions on Crowdin. If you are the contributor of any of the added ones, or participated in the updates of any, please, submit a PR adding your proper credit as the "README.md" suggests. Also removes translations that are incomplete at the moment of this commit. The credits for those will not be removed.