fix: root-related mounts leak when using with Magisk
This commit fixes an issue where using ReZygisk together with Magisk could lead to root-related mounts being leaked, when using modules that mount to partitions such as "/system". This happens because Magisk sets the source of some of its mounts to "worker", which wasn't previously recognized. With that it also removes the whitelist for some "/system" mounts to not be umounted.