fix: preload on some systems; update: rollback global module `on_load`
This commit fixes the preload on some systems by opening modules before system server is forked. For that, we need to handle the solist early on to avoid gaps due to libraries being opened after that. And, because of that, we also deal with the deconstructors manually. It also rolls back the commit that made module "on load" be called in preload stage/globally, as initializators can be used instead, and makes it behave the same way as other Zygisks.